requirements.txt
Pillow>=5.0
qrcode>=6.0
